snow_rest/requirements.txt
snowflake-snowpark-python
flask
